// client/src/features/submission/models/submission-status.ts
export enum SubmissionStatus {
  IN_QUEUE = "In Queue",
  PROCESSING = "Processing",
  ACCEPTED = "Accepted",
  WRONG_ANSWER = "Wrong Answer",
  TIME_LIMIT_EXCEEDED = "Time Limit Exceeded",
  COMPILATION_ERROR = "Compilation Error",
  INTERNAL_ERROR = "Internal Error",
  EXEC_FORMAT_ERROR = "Exec Format Error",
  RUNTIME_ERROR = "Runtime Error",
  POLLING = "Polling",
  POLLING_ERROR = "Polling Error",
}